seperated payment methods
